remove parsing from template
The aoc_utils crate replaced nom with winnow for parsing.
It would've been an option to translate the template accordingly.
But a parsing library is overkill for most puzzles and I'm not
yet comfortable enough with winnow to make the template idiomatic.

winnow-based parsing may be added back to the template at some point.